requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-huggingface==0.0.3

transformers
sentence-transformers
faiss-cpu
torch
python-dotenv
pypdf